cmake: Fix race condition in CMakeLists.txt
The "glad" target depends on the generated source files. It may happen happen
that these are not yet generated in a multithreaded build, so we make the
dependency explicit here. There's a new target called "glad-generate-files"
that requires the files to exist. The "glad" target depends on the new target.