CMake support for generating all extensions
Add a new option GLAD_ALL_EXTENSIONS, OFF by default. When turned ON, the value of the variable GLAD_EXTENSIONS is not passed on to the python code generator. As a result, the generated code includes all extensions in the requested APIs.

This is needed because there does not appear to be a way to distinguish an unset variable from one with an empty value. An empty GLAD_EXTENSIONS value indicates that the user does not want any extensions to be included.