c: remove _gl_handle
This static pointer serves no purpose whatsoever. In the non-ondemand
case, it never becomes relevant because the only codepath that accesses
it is `gladLoaderLoadGL[Context]`, which immediately unloads the handle
again before returning.

In the ondemand case, it is fully redundant with the `handle` field that
already exists in the internal global userptr. So there is no point in
storing the handle separately in this case either.