c: remove _egl_handle
Similar justification to the previous commit. Something interesting to
note is that, out of the box, this code differs from the GL case in that
the handle is actually kept around until the context is unloaded, even
though again nothing uses it. Change this behavior, to make it
consistent with how the GL loader works.