c: store mx-global context as a pointer
Instead of a separate struct that needs to be copied by-value. This will
reduce the risk of accidentally ending up with an inconsistent context
state.

We still need a default context in the even that a user loads
an `mx-global` context without ever supplying a custom struct, but this
is easy enough to return in an ad-hoc fashion whenever `gladGet*Context`
is called without a context otherwise being bound.

(Incidentally, the way this logic is implemented also lets you recover
the global context by explicitly calling `gladSet*Context(NULL)`)